haulage_system/requirements.txt
asgiref==3.8.1
brotli==1.2.0
cffi==2.0.0
cssselect2==0.8.0
Django==5.1.6
djangorestframework==3.17.1
fonttools==4.61.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
python-decouple==3.8
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.1
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
drf-yasg
django-cors-headers